Much cleaning, many refactor. Such wow
