+++ knex/knexfile.ts
@@ -1,58 +1,48 @@
-// // Update with your config settings.
-
-// module.exports = {
-
-//   development: {
-//     client: "sqlite3",
-//     connection: {
-//       filename: "./dev.sqlite3"
-//     }
-//   },
-
-//   staging: {
-//     client: "postgresql",
-//     connection: {
-//       database: "my_db",
-//       user: "username",
-//       password: "password"
-//     },
-//     pool: {
-//       min: 2,
-//       max: 10
-//     },
-//     migrations: {
-//       tableName: "knex_migrations"
-//     }
-//   },
-
-//   production: {
-//     client: "postgresql",
-//     connection: {
-//       database: "my_db",
-//       user: "username",
-//       password: "password"
-//     },
-//     pool: {
-//       min: 2,
-//       max: 10
-//     },
-//     migrations: {
-//       tableName: "knex_migrations"
-//     }
-//   }
-
-// };
-
 export = {
   development: {
-    client: 'pg',
-    connection: process.env.DATABASE_URL,
+    client: 'sqlite3',
+    connection: {
+      filename: './dev.sqlite3.db',
+    },
+    useNullAsDefault: true,
+    migrations: {
+      directory: __dirname + '/migrations/dev',
+    },
     seeds: {
-      directory: './seeds/dev',
+      directory: __dirname + '/seeds/dev',
     },
   },
+
+  staging: {
+    client: 'pg',
+    connection: process.env.DATABASE_URL,
+
+    pool: {
+      min: 2,
+      max: 10,
+    },
+
+    migrations: {
+      directory: __dirname + '/migrations/dev',
+    },
+
+    seeds: {
+      directory: __dirname + '/seeds/dev',
+    },
+  },
+
   production: {
     client: 'pg',
     connection: process.env.DATABASE_URL,
+
+    pool: {
+      min: 2,
+      max: 10,
+    },
+
+    migrations: {
+      directory: __dirname + '/migrations/prod',
+    },
+
   },
 };
