+++ package.json
@@ -46,6 +46,7 @@
     "@types/uuid": "^3.4.7",
     "dotenv": "^8.2.0",
     "nodemon": "^1.19.4",
+    "sqlite3": "^4.1.1",
     "ts-node": "^8.6.2",
     "tslint": "^6.0.0",
     "typescript": "^3.8.2"
