+++ src/middlewares/authmw.ts
@@ -1,8 +1,7 @@
 /* eslint-disable no-underscore-dangle */
 import { Response, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
-import User from '../db/controllers/users';
-import { QueryResult } from 'pg';
+import User from '../models/user';
 import { RequestWithUser } from '../types/request';
 
 interface Token {
@@ -39,9 +38,9 @@ const authmw = {
              *    so if someone had tampered with the payload, the jwt verification
              *    would fail anyway
              */
-            const { rows } = await User.getOne(decoded.userId) as QueryResult;
+            const user = await User.query().where({ id: decoded.userId });
 
-            if (!rows[0]) {
+            if (!user[0]) {
               res.status(403).end({ message: 'Access denied' });
             } else {
               req.user = { id: decoded.userId };
